test/requirements.txt
bson==0.5.10
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.0.12
cryptography==39.0.0
idna==3.4
pycparser==2.21
PyJWT==2.4.0
pymongo==4.1.1
python-dateutil==2.8.2
requests==2.27.1
six==1.16.0
urllib3==1.26.12
